+++ R/Spectra.R
@@ -21,20 +21,23 @@
 #'
 #' The `<type of data>` allows to define which data should be extracted from
 #' the selected spectra. MassQL defines *type of data* being `MS1DATA` or
-#' `MS2DATA` to retrieve MS1 or MS2 scans. In addition, functions can be applied
-#' to these to modify (e.g. sum) the data. `SpectraQL` supports:
+#' `MS2DATA` to retrieve data from MS1 or MS2 scans. By default peak data will
+#' be returned, but in addition, MASSQL defines additional functions that can
+#' be applied to modify the data or select different data to be returned. In
+#' addition `SpectraQL` defines the special type of data `*` which will return
+#' the results as a `Spectra` object. `SpectraQL` supports:
 #'
 #' - `*`: select all data and return the data subset as a [Spectra()] object.
-#' - `MS1DATA`: return a [Spectra()] with all MS1 scans from the selected
-#'   spectra.
-#' - `MS2DATA`: return a [Spectra()] with all MS2 scans from the selected
-#'   spectra.
+#' - `MS1DATA`: return the [peaksData()] from all selected **MS1** spectra,
+#'   i.e. a `list` with two column matrices with the peaks' m/z and intensity
+#'   values.
+#' - `MS2DATA`: return the [peaksData()] from all selected **MS2** spectra,
+#'   i.e. a `list` with two column matrices with the peaks' m/z and intensity
+#'   values.
 #' - `scaninfo(MS1DATA)`, `scaninfo(MS2DATA)`: return the [spectraData()] of all
 #'   selected spectra.
 #' - `scansum(MS1DATA)`, `scansum(MS2DATA)`: sum of the peak intensities of
 #'   the selected spectra (TIC).
-#' - `scannum(MS1DATA)`, `scannum(MS2DATA)`: the scan number(s) of the selected
-#'   spectra.
 #'
 #' @section Conditions:
 #'
@@ -52,7 +55,7 @@
 #' - `CHARGE`: the charge for MS2 spectra.
 #' - `POLARITY`: the polarity of the spectra (can be `"positive"`, `"negative"`,
 #'   `"pos"` or `"neg"`, case insensitive).
-#' - `MS2PROD`: allows to select MS2 spectra that contain a peak with
+#' - `MS2PROD` or MS2MZ`: allows to select MS2 spectra that contain a peak with
 #'   particular m/z value(s). See below for examples.
 #' - `MS2PREC`: allows to select MS2 spectra with the defined precursor m/z
 #'   value(s). See below for examples.
@@ -85,6 +88,8 @@
 #'
 #' @importClassesFrom Spectra Spectra
 #'
+#' @importFrom methods is
+#'
 #' @exportMethod query
 #'
 #' @examples
@@ -99,9 +104,11 @@
 #' ## Subset to spectra measured between 300 and 400 seconds
 #' query(sps_dda, "QUERY * WHERE RTMIN = 300 AND RTMAX = 400")
 #'
-#' ## To select only MS1 or MS2 spectra use "MS1DATA" or "MS2DATA" instead
-#' ## or *. Note also that queries are case-insensitive.
-#' query(sps_dda, "query ms1data where rtmin = 300 and rtmax = 400")
+#' ## To extract peaks data from MS1 or MS2 spectra use "MS1DATA" or "MS2DATA"
+#' ## instead of *. Note also that queries are case-insensitive.
+#' pks <- query(sps_dda, "query ms1data where rtmin = 300 and rtmax = 400")
+#' pks
+#' head(pks[[1L]])
 #'
 #' ## To select (MS2) spectra with a certain precursor m/z the MS2PREC condition
 #' ## can be used. Below we extract all spectra with a precursor m/z of 99.9
@@ -122,6 +129,9 @@
 #' ## In contrast, do select MS2 spectra containing a peak with a certain m/z
 #' ## we have to use the condition MS2PROD
 #' query(sps_dda, "QUERY * WHERE MS2PROD = 100:TOLERANCEMZ=1")
+#'
+#' ## MS2MZ can be used as alternative to MS2PROD
+#' query(sps_dda, "QUERY * WHERE MS2MZ = 100:TOLERANCEMZ=1")
 setMethod("query", "Spectra", function(x, query = character(), ...) {
     if (!length(query))
         return(x)